[core][presenters] remove: perv and wipe presenters
